% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_interactions.R
\name{iv_2x2}
\alias{iv_2x2}
\title{IV Estimation of 2x2 Factorial Design}
\usage{
iv_2x2(y, d, z, max_iter = 15, tol = 1e-06)
}
\arguments{
\item{y}{vector of outcomes.}

\item{d}{2-column matrix of binary treatments.}

\item{z}{2-column matrix of binary instruments.}

\item{max_iter}{maximum number of iterations to conduct when
estimating the GMM.}

\item{tol}{criteria for stopping the iterative GMM procedure.}
}
\value{
A list of class \code{iv2x2} that contains the following
  components: 
\item{rho}{vector of estimated compliance class
  probabilities.}
\item{psi}{vector of the estimated conditional mean of the outcome
  within the compliance classes.}
\item{init}{list of two vectors, with the initial starting values
  for the GMM estimation based on a just-identified estimation.}
\item{vcov}{estimated asymptotic variance matrix of the combined
  \code{rho} and \code{psi} parameters.}
\item{tau}{vector of estimated causal effects within levels of the
  compliance classes.}
\item{tau_se}{vector of estimated standard errors for the
  estimated effects in \code{tau}.}
}
\description{
Estimates principal stratum-specific effects and interactions in a
2x2 factorial experiment
}
\details{
This function estimates treatment effects for 2x2 factorial
experiments in the face of noncompliance on both factors. A
monotonicity assumption is assumed for both treatment-instrument
pairs, along with treatment exclusion. See Blackwell (2017) for
more details on those assumptions.

The procedure uses iterative generalized method of moments (GMM)
to  estimate both the proportions of each compliance class (also
known as principal strata) and the average potential outcomes
within those classes. It also provides estimates of several
one-way, joint, and interactive treatment effects within these
classes. 

Under the above assumptions, the compliance classes are the
product of the compliance classes for each treatment-instrument
pair. For instance, \code{"cc"} is the class that would comply
with both treatments, \code{"ca"} is the class that would comply
with the first treatment and always take the second treatment, and
\code{"cn"} is the class that would comply with the first
treatment and never take the second treatment. In the estimated
effects, some effects are defined for the \code{"c-"} or
\code{"-c"} compliance classes, which are effects for compliers on
one treatment, averaging over the compliance status of the other
treatment. Finally, note that treatment effects are only
well-defined for compliance classes for which there is compliance
on at least one treatment.
}
\references{
Matthew Blackwell (2017) Instrumental Variable Methods
  for Conditional Effects and Causal Interaction in Voter
  Mobilization Experiments, Journal of the American Statistical
  Association, 112:518, 590-599,
  \url{https://doi.org/10.1080/01621459.2016.1246363}
}
\author{
Matt Blackwell
}
