% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary}
\alias{primary}
\title{Data on the direct primary in US congressional elections}
\format{
A data frame with 1164 observations and the following 7
  variables:
\describe{
  \item{state}{name of the state}
  \item{year}{year of the congressional election}
  \item{dem_share}{percentage of the total vote cast for the
  Democratic candidate, 0-100}
  \item{rep_share}{percentage of the total vote cast for the
  Republican candidate, 0-100}
  \item{other_share}{percentage of the total vote cast for other
  parties, 0-100}
  \item{primary}{binary variable indicating if the state had the
  direct primary (=1) or not (=0)}
  \item{south}{binary variable indicating if the state is in the
  South (=1) or not (=0)}
}
}
\source{
\url{https://www.icpsr.umich.edu/icpsrweb/ICPSR/studies/6895}
}
\usage{
primary
}
\description{
A data set on the presence of the direct primary in U.S.
congressional elections and the vote shares for the Democratic,
Republican, and third parties. Based on ICPSR Study 6985
}
\references{
David, Paul T., and Claggett, William. Party Strength
  in the United States: 1872-1996. Ann Arbor, MI: Inter-university
  Consortium for Political and Social Research [distributor],
  2008-09-10.  https://doi.org/10.3886/ICPSR06895.v1
}
\keyword{datasets}
