% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remit}
\alias{remit}
\title{Cross-national data on remittances and protest}
\format{A data frame with 2429 observations and 14 variables.}
\source{
\url{https://doi.org/10.7910/DVN/TVZQG6}
}
\description{
A data set to replicate the findings of Escrib\`{a}-Folch,
Meseguer, and Wright (2018)
}
\details{
\itemize{
 \item Protest.
 \item remit.
 \item dict.
 \item l1gdp.
 \item l1pop.
 \item l1nbr5.
 \item l12gr.
 \item l1migr.
 \item elec3.
 \item cowcode.
 \item period.
 \item distwremit.
 \item caseid.
 \item year.
}
}
\references{
Escrib\`{a}-Folch, A., Meseguer, C. and Wright, J. (2018), Remittances and Protest in Dictatorships. American Journal of Political Science, 62: 889-904. doi:10.1111/ajps.12382

Wright, Joseph, 2018, "Replication Data for: Remittances and Protest in Dictatorships", https://doi.org/10.7910/DVN/TVZQG6, Harvard Dataverse, V1, UNF:6:IE6OqUb3EB5AIDYKI28mgA== [fileUNF]
}
